Fix tmux reboot detect bug & improve code style
tmux-status-right was not correctly detecting the need for a reboot due
to overwriting some vars & setting the status style in the .tmux.conf
file which prevents the status-right script from changing it.

The script also wasn't following my bash coding style conventions so it
has been updated to follow it and also be easier to read and follow.