Improve 0.8 nozzle settings & update Prusa config
Lots of small adjustments to the 0.8mm nozzle configs such as adding an
ultradetail (beta/untested) config and better settings for the 0.67mm
layer height settings.

Also updated the PrusaResearch config ini to the latest snapshot from
the upstream repo.

One more tweak is that I renamed "Diamondback" to just "Diamond" for the
sake of brevity in config names.