Add submodule check to stow.sh and ensure it doesn't crash shell
Add a check to ensure submodules are loaded before trying to stow
anything. Dotfiles behave strangely if submodules were not loaded.

Requiring stow.sh to be sourced can lead to the entire shell exiting
if a command failed thanks to the `set -e` flag. This flag should never
be set in a script that is meant to be sourced.

Also, setting the `set -e` flag in a sourced file can also lead to any
subsequent scripts closing the shell if `set +e` is not added in the
same script.