Add systemd checks for all users & show root mail status in tmux
All system users with the "linger" flag enabled will have their status
checked to show it in the tmux status bar.

Root mail will also be checked to see if there is any mail.

Both of these features assume that the user's sudoers file has NOPASSWD
for the commands the script uses.